+++ src/js/moveSymbol.js
@@ -335,12 +335,7 @@ const drop = (event) => {
   target.classList.add('draggable');
   target.dataset.symbolInfo = data;
   target.innerHTML = document.querySelector('.newSVG > svg').innerHTML;
-  //! BUG: Cannot drop symbols on top of grid lines. Try a try-catch block and adjust the symbols coordinates
-  try {
-    event.target.offsetParent.appendChild(target);
-  } catch (error) {
-    target.setAttribute('viewBox', '25, 50, 150, 100');
-  }
+  event.target.offsetParent.appendChild(target);
   // Get the BBox only after the target has been appended
   const bbox = target.getBBox();
   // The bounceIn animation fucks everything up, remove it and set the BBox of the parent draggable symbol